+++ src/main/java/com/example/index/generators/HypernymIndexGenerator.java
@@ -111,7 +111,7 @@ public final class HypernymIndexGenerator extends IndexGenerator<DependencyEntry
                     }
 
                     batch.add(new DependencyEntry(
-                        rs.getInt("dependency_id"),
+                        rs.getLong("dependency_id"),
                         rs.getInt("document_id"),
                         rs.getInt("sentence_id"),
                         rs.getInt("begin_char"),
