+++ src/main/java/com/example/index/generators/POSIndexGenerator.java
@@ -72,7 +72,7 @@ public final class POSIndexGenerator extends IndexGenerator<AnnotationEntry> {
             try (ResultSet rs = stmt.executeQuery()) {
                 while (rs.next()) {
                     AnnotationEntry entry = new AnnotationEntry(
-                        rs.getInt("annotation_id"),
+                        rs.getLong("annotation_id"),
                         rs.getInt("document_id"),
                         rs.getInt("sentence_id"),
                         rs.getInt("begin_char"),
